Fix #13 - Add Fallback for Node.js Versions lower than 10.x
`npm test` now fails when running at this commit, but
only with node versions < 10.x. This needs to be addressed. In the
worst case we just make the test pass when such a version is used.
It's not a a nice solution, but I didn't find any documentation on
how to make C++ Addons work when re-requiring them after clearing
the require-cache so far. I will keep looking though :)